Introduce changes from WHMCS 7.0 to module CORE-10577
This commit introduces some of the changes to the plesk
module WHMCS developed as part of their feature work in the 7.0
series. The biggest change comes from the introduction of simple
mode configuration. To aid this we implemented a new API function
and this including the matching API template. It also contains
some hygiene fixes: adding use statements, migrating from full class
names to the short names based on use statements for input sanitization
which set you up for namespace support, and a few parameter tags and
checks, which allow the IDE to do better static analysis.

Merging this commit to the master branch ensures that the code in
your repo is the same as the code in the WHMCS mainline branch and
should reduce the effort necessary to pull in changes in future
releases.