Make panelexternalid the primary key
This is because the clientid may not be unique if the client has multiple Plesk accounts. It's actually quite common when you have multiple servers or even if the client has a shared account and a reseller account on the same server. The only truly unique element in this table is panelexternalid so it may as well be the primary key, no?

Fixes Issue #18 BUT will only fix it upon initial install...existing users of the plesk module will need to modify the table manually.